Update Wifi standard based on band and mode capability.
In scan window of wifi it is showing wifi5 when AP
operating in 11AX 2Ghz and Device capability is 11AC,
whereas after connection to 11ax 2Ghz AP it is showing
wifi4 standard. Thus scan and connection windows are not
in sync. This commit adds fix to show wifi4 standard in
both scan and connection windows when device max capability
supported is 11AC and AP operating 11AX 2Ghz band.

CRs-Fixed: 3010519
Change-Id: I42895e043c2d4784ee58049bc6d6da6aaf80ac5d