STA: Reset mTargetNetworkId based on new connection status
Currently mTargetNetworkId is reset in disconnected state
to allow subsequent connections. During connection to new
AP framework receive NETWORK_DISCONNECT_EVENT to old AP
from supplicant due to which mTargetNetworkId is reset,
but when AUTHENTICATION_FAILURE_EVENT to new AP received
in framework inorder to update network selection status
mTargetNetworkId is needed which is cleared during the
disconnect event.

Added changes to not clear mTargetNetworkId during new
connection.

Currently ssid is not sent as part of disconnect event
to framework.

Added changes to include ssid in disconnect event which
is used to determine the status of new connection.

NOT A CONTRIBUTION

CRs-Fixed: 3120243
Change-Id: Ibf3f685480229b48c7f332a6557bb14e79e870a3